requirements.txt
Babel==1.3
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-Table==0.2.4
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.9.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
coverage==3.7.1
decorator==3.4.0
ecdsa==0.11
ipython==1.2.0
itsdangerous==0.24
mock==1.0.1
pbr==0.10.0
pycrypto==2.6.1
pytz==2014.7
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.12
stevedore==0.15
wsgiref==0.1.2
